refactor(web): refactor-item-information-card
